Remove non-tests from the tests folder
